move all cps files
