+++ taxdata/cps/impute.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import numpy as np
 from scipy.stats import norm
-from helpers import log, LOG_VARS, X_VARS, filingparams, cps_yr_idx
+from .helpers import log, LOG_VARS, X_VARS, filingparams, cps_yr_idx
 
 
 def impute(df, logit_betas, ols_betas, x_vars, l_adj, o_adj, prob_mult):
