+++ taxdata/cps/transform_sas.py
@@ -4,7 +4,7 @@ to Python
 """
 import pickle
 from pathlib import Path
-from cps_meta import CPS_META_DATA
+from .cps_meta import CPS_META_DATA
 
 
 CUR_PATH = Path(__file__).resolve().parent
